requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
aiohttp==3.9.1
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
numpy==1.26.2
scikit-learn==1.3.2
tenacity==8.2.3
prometheus-client==0.19.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
pyyaml==6.0.1
Jinja2==3.1.2